requirements.txt
spotipy==2.19.0
#deemix==3.6.6
deezer-py==1.3.6
pyfiglet==0.7.5
chump==1.6.0
GitPython==3.1.12
requests==2.25.1
tabulate==0.8.7
psutil==5.9.0
git+https://github.com/jbh-cloud/spotify_download_deemix@v3.6.7#egg=deemix
jsonschema~=4.4.0